+++ clients/web3-wallet/src/rest-api/modules/web-socket-api.ts
@@ -25,7 +25,7 @@ import type { GetWebSocketAuthTokenResponse } from '../types';
 const WebSocketApiAxiosParamCreator = function (configuration: ConfigurationRestAPI) {
     return {
         /**
-         * Get a WebSocket auth token before connecting to the WebSocket stream service.
+         * Get a WebSocket auth token before connecting to the WebSocket stream service. For details on WebSocket usage, please refer to the [WebSocket documentation](/products/websocket-api/introduction).
          *
          *
          * @summary Get WebSocket Auth Token
@@ -70,7 +70,7 @@ const WebSocketApiAxiosParamCreator = function (configuration: ConfigurationRest
  */
 export interface WebSocketApiInterface {
     /**
-     * Get a WebSocket auth token before connecting to the WebSocket stream service.
+     * Get a WebSocket auth token before connecting to the WebSocket stream service. For details on WebSocket usage, please refer to the [WebSocket documentation](/products/websocket-api/introduction).
      *
      *
      * @summary Get WebSocket Auth Token
@@ -118,7 +118,7 @@ export class WebSocketApi implements WebSocketApiInterface {
     }
 
     /**
-     * Get a WebSocket auth token before connecting to the WebSocket stream service.
+     * Get a WebSocket auth token before connecting to the WebSocket stream service. For details on WebSocket usage, please refer to the [WebSocket documentation](/products/websocket-api/introduction).
      *
      *
      * @summary Get WebSocket Auth Token