+++ clients/web3-wallet/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 12.2.0 - 2026-09-10
+
+### Changed (2)
+
+- Added parameter `enableRFQ`
+  - affected methods:
+    - `quoteAndBuildSwapTransaction()` (`GET /api/v1/dex/aggregator/quote-and-swap`)
+- Added parameter `excludeDexes`
+  - affected methods:
+    - `quoteAndBuildSwapTransaction()` (`GET /api/v1/dex/aggregator/quote-and-swap`)
+
 ## 12.1.2 - 2026-09-03
 
 ### Changed (2)